Removes restriction on lower and upper bounds of ra/dec query parameters.
Previously, the function restricted queries for RA/DEC values outside
of the ranges 0,24 and -90,90, for RA and DEC, respectively.

Lifting this restriction allows for finding data with RA/DEC values
outside of this range.

This is useful because SETI sets the RA/DEC values
when observing objects that are not fixed to particular celestial
coordinates, such as spacecraft, to values outside of the physical
range. (Spacecraft are at RA,DEC = -99,-99.)

However, the default values, when no values are provided in the
query are:
ramin = 0, ramax = 24, decmin = -90, decmax = 90